+++ fetch_links.py
@@ -17,8 +17,7 @@ from logging.handlers import RotatingFileHandler
 from pathlib import Path
 
 # Custom libraries
-# import twitter_links
-import twitter_links_v2
+import twitter_links
 import reddit_links
 import rss_links
 import startup_and_validate
@@ -40,7 +39,7 @@ def fetch_links(config, sources):
     # TODO: Spin up threads to run these in parallel
     # rss_links.run(sources['rss']['feeds'], config['db_info'])
     # reddit_links.run(sources['reddit'], config['db_info'])
-    twitter_links_v2.run(sources['twitter'], config['db_info'])
+    twitter_links.run(sources['twitter'], config['db_info'])
 
 
 def main():
